simplify training page
